Fix an assertion failure
`Id::operator-=(a,b)` throws when a == b, because you're not supposed to
make an invalid Id from a valid Id. So instead we use `uint64_t` to
represent the page offsets, because those are not necesarily valid Ids.

This assertion triggered when I tried to add units to the Haskell
indexer, I'm not sure why it hasn't triggered before.