Support multiple output files
We had "groups" as a way to run a test multiple ways where the output
files might differ, but that only works if you can run the test multiple
ways without recompiling. For tests where the output depends on
something in the environment, we need a way to manage multiple output
files, so that's what this does.

Additionally I've used it in the Haskell indexer tests, where the GHC
version affects the output.