[glass] Optionally return full file content
Clients may want to update symbol positions to account for
local changes to the source file, and for this they need the original
file content as it was indexed. Currently the way to do that is to use
the `revision` returned by Glass and retrieve the file from source
control, but that might not be possible in general:

 * the indexed file might not be checked in (consider interactively
   indexing code in the background when using an IDE), or
 * the files might not be in source control at all (consider indexing
   packages you downloaded from a package manager).

This provides a way for the Glass client to ask Glass for the original
file content that was indexed. The content might not be available - it
relies on the indexer having produced the appropriate `src.FileContent`
facts, and only the Haskell indexer does this right now, but it could
become more widely supported in the future.

We could also have Glass try to get the content from the SCM; I haven't
implemented that.