Move to 64-bit Unit and Ownership IDs
Summary:
# Ownership and Unit identifier bump to 64-bit
We're reaching the limit of 32-bit identifiers for our larger repositories.
Ownership calcluations rely on the native 64-bit fact IDs being in the 32-bit range. Unit IDs are also stored in 32-bits.

This diff increases the size of ownership sets IDs and Unit IDs to 64-bits to allow for code growth.

There's an estimated 2% increase in DB size - ~~this needs to be validated on actual indexed DBs.~~
This has been validated at less than 2% on the `www.hack.light` db. Details [here](https://docs.google.com/document/d/1Kv76vKJ9MDHYysVqiuhsvHFxeUxBl69P0c95Qq3uP_k/edit?tab=t.0)

# Stack
Due to the size of the change and number of files touched, the change is split into a stack of 4 diffs. This is diff (1/4)

Reviewed By: iamirzhan

Differential Revision: D88004735

fbshipit-source-id: 539f4e28f6e8924c09935ba653b51010c58bcfde