[rts] Seek iterators don't need to return facts in lexicographic order
I want to relax the restriction that seek() must create an iterator that
returns facts in lexicographic order, because this gives more
flexibility in the storage layer. In particular we don't need to sort
all the keys, which can be problematic since keys can be large.

It turns out that we don't rely on the lexicographic ordering except for
saving and resuming query continuations, and this can be done by using
the current fact as the checkpoint instead.

We can remove the merge() operation on FactIterator and use only
append(), and this allows some optimisations when resuming iterators.